Implement transition.loop( targetObject, transitionParams ) for transitions that start and end in the same point
Similar to what easing.continuousLoop does, where this:

  local rect = display.newRect(200, 200, 50, 50)
  transition.loop(rect, {
    time         = 2000,
    y               = 100,
    iterations = 20,
    transition = easing.inOutQuad
  })

Is gonna loop a rectangle floating up and down 20 times before ending in the same position it started, the whole animation (up + down) takes 2 secs, but it also uses easing.inOutQuad for both moves.

I face my self a lot of times implementing and in and out animations for lot of stuff like blinking (fadeIn -> fadeOut), floating, etc. with this I just need a single transition and the result is even smoother and more precise.